test: Comment, Poll 관련 도메인 파일 위치 이동
