+++ backend/codelight/src/main/java/com/project/codelight/comment/domain/CommentLike.java
@@ -1,4 +1,4 @@
-package com.project.codelight.post.domain;
+package com.project.codelight.comment.domain;
 
 import static lombok.AccessLevel.PROTECTED;
 
@@ -9,10 +9,12 @@ import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.Index;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import jakarta.persistence.UniqueConstraint;
+import java.time.LocalDateTime;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -20,32 +22,31 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = PROTECTED)
 @Entity
-@Table(name = "poll_votes",
-    uniqueConstraints = @UniqueConstraint(name = "uk_poll_option_user", columnNames = {"poll_id",
-        "option_id", "user_id"}))
-public class PollVote {
+@Table(name = "comment_likes",
+    uniqueConstraints = @UniqueConstraint(name = "uk_comment_user", columnNames = {"comment_id",
+        "user_id"}),
+    indexes = {@Index(name = "idx_user_id", columnList = "user_id")})
+public class CommentLike {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "vote_id")
+    @Column(name = "like_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "poll_id", nullable = false)
-    private Poll poll;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "option_id", nullable = false)
-    private PollOption option;
+    @JoinColumn(name = "comment_id", nullable = false)
+    private Comment comment;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
     private User user;
 
+    @Column(name = "created_at", nullable = false, updatable = false, insertable = false)
+    private LocalDateTime createdAt;
+
     @Builder
-    private PollVote(Poll poll, PollOption option, User user) {
-        this.poll = poll;
-        this.option = option;
+    private CommentLike(Comment comment, User user) {
+        this.comment = comment;
         this.user = user;
     }
 }
