+++ backend/codelight/src/main/java/com/project/codelight/comment/repository/CommentRepository.java
@@ -1,6 +1,6 @@
 package com.project.codelight.comment.repository;
 
-import com.project.codelight.post.domain.Comment;
+import com.project.codelight.comment.domain.Comment;
 import com.project.codelight.post.domain.Post;
 import com.project.codelight.user.domain.User;
 import java.util.List;
