+++ backend/codelight/src/main/java/com/project/codelight/comment/service/CommentService.java
@@ -1,9 +1,9 @@
 package com.project.codelight.comment.service;
 
+import com.project.codelight.comment.domain.Comment;
 import com.project.codelight.comment.repository.CommentRepository;
 import com.project.codelight.global.exception.CodeLightException;
 import com.project.codelight.global.exception.ExceptionCodeType;
-import com.project.codelight.post.domain.Comment;
 import com.project.codelight.post.domain.Post;
 import com.project.codelight.post.repository.PostRepository;
 import com.project.codelight.user.domain.User;
