+++ backend/codelight/src/main/java/com/project/codelight/poll/controller/PollController.java
@@ -3,11 +3,11 @@ package com.project.codelight.poll.controller;
 import com.project.codelight.auth.security.model.CustomUserDetails;
 import com.project.codelight.global.exception.CodeLightException;
 import com.project.codelight.global.exception.ExceptionCodeType;
+import com.project.codelight.poll.domain.Poll;
+import com.project.codelight.poll.domain.PollOption;
 import com.project.codelight.poll.dto.request.PollVoteRequest;
 import com.project.codelight.poll.dto.response.PollResponse;
 import com.project.codelight.poll.service.PollService;
-import com.project.codelight.post.domain.Poll;
-import com.project.codelight.post.domain.PollOption;
 import com.project.codelight.user.domain.User;
 import jakarta.validation.Valid;
 import java.util.List;
