+++ backend/codelight/src/main/java/com/project/codelight/poll/domain/PollOption.java
@@ -1,4 +1,4 @@
-package com.project.codelight.post.domain;
+package com.project.codelight.poll.domain;
 
 import static lombok.AccessLevel.PROTECTED;
