+++ backend/codelight/src/main/java/com/project/codelight/poll/domain/PollVote.java
@@ -1,4 +1,4 @@
-package com.project.codelight.post.domain;
+package com.project.codelight.poll.domain;
 
 import static lombok.AccessLevel.PROTECTED;
 
@@ -9,12 +9,10 @@ import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.Index;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import jakarta.persistence.UniqueConstraint;
-import java.time.LocalDateTime;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -22,31 +20,32 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = PROTECTED)
 @Entity
-@Table(name = "comment_likes",
-    uniqueConstraints = @UniqueConstraint(name = "uk_comment_user", columnNames = {"comment_id",
-        "user_id"}),
-    indexes = {@Index(name = "idx_user_id", columnList = "user_id")})
-public class CommentLike {
+@Table(name = "poll_votes",
+    uniqueConstraints = @UniqueConstraint(name = "uk_poll_option_user", columnNames = {"poll_id",
+        "option_id", "user_id"}))
+public class PollVote {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "like_id")
+    @Column(name = "vote_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "comment_id", nullable = false)
-    private Comment comment;
+    @JoinColumn(name = "poll_id", nullable = false)
+    private Poll poll;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "option_id", nullable = false)
+    private PollOption option;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
     private User user;
 
-    @Column(name = "created_at", nullable = false, updatable = false, insertable = false)
-    private LocalDateTime createdAt;
-
     @Builder
-    private CommentLike(Comment comment, User user) {
-        this.comment = comment;
+    private PollVote(Poll poll, PollOption option, User user) {
+        this.poll = poll;
+        this.option = option;
         this.user = user;
     }
 }
